% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_scale.R
\name{change_scale}
\alias{change_scale}
\alias{change_scale.numeric}
\alias{change_scale.grouped_df}
\alias{change_scale.data.frame}
\title{Rescale a numeric variable}
\usage{
change_scale(x, ...)

\method{change_scale}{numeric}(x, to = c(0, 100), range = NULL,
  verbose = TRUE, ...)

\method{change_scale}{grouped_df}(x, select = NULL, exclude = NULL,
  to = c(0, 100), range = NULL, ...)

\method{change_scale}{data.frame}(x, select = NULL, exclude = NULL,
  to = c(0, 100), range = NULL, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Arguments passed to or from other methods.}

\item{to}{New range of values of the data after rescaling.}

\item{range}{Initial (old) range of values. If \code{NULL}, will take the range of data.}

\item{verbose}{Toggle warnings on or off.}

\item{select}{For a data frame, character vector of column names to be
standardized. If \code{NULL} (the default), all variables will be
standardized.}

\item{exclude}{For a data frame, character vector of column names to
be excluded from standardization.}
}
\value{
A rescaled object.
}
\description{
Rescale a numeric variable. This scales all numeric variables in the range 0 - 1.
}
\examples{
change_scale(c(0, 1, 5, -5, -2))
change_scale(c(0, 1, 5, -5, -2), to = c(-5, 5))

head(normalize(iris))

}
\seealso{
\code{\link{normalize}} \code{\link{standardize}} \code{\link{ranktransform}}
}
