% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_teststats_to_effsize.R
\name{chisq_to_phi}
\alias{chisq_to_phi}
\alias{chisq_to_cramers_V}
\title{Compute Effect sizes for Contingency Tables (\eqn{chi^2})}
\usage{
chisq_to_phi(chisq, N)

chisq_to_cramers_V(chisq, N, a, b)
}
\arguments{
\item{chisq}{The \eqn{chi^2} statistic.}

\item{N}{The sample size}

\item{a}{The number of rows in the contingency table.}

\item{b}{The number of columns in the contingency tables.}
}
\value{
A numeric integer between 0-1.
}
\description{
These convert \eqn{chi^2} test statistics to \eqn{\phi} and \eqn{\Cramer's V}.
}
\examples{
\dontrun{
ctab <- as.table(rbind(c(762, 327, 468), c(484, 239, 477), c(484, 239, 477)))
N <- sum(ctab)
a <- nrow(ctab)
b <- ncol(ctab)

(Xsq <- chisq.test(ctab))
#
#         Pearson's Chi-squared test
#
# data:  ctab
# X-squared = 41.234, df = 4, p-value = 2.405e-08

chisq_to_phi(30.07, N)
chisq_to_cramers_V(30.07, N, a, b)
}

}
