% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_posteriors.R
\name{convert_posteriors_to_r}
\alias{convert_posteriors_to_r}
\alias{posteriors_to_r}
\alias{convert_posteriors_to_t}
\alias{posteriors_to_t}
\title{Convert posterior distributions from a Bayesian model}
\usage{
convert_posteriors_to_r(model, ...)

posteriors_to_r(model, ...)

convert_posteriors_to_t(model, ...)

posteriors_to_t(model, ...)
}
\arguments{
\item{model}{A Bayesian statistical model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Convert posterior distributions from a Bayesian model to indices of effect size.
}
\examples{
library(rstanarm)
model <- stan_glm(mpg ~ cyl, data = mtcars, refresh = 0, chains = 2)

convert_posteriors_to_t(model)
convert_posteriors_to_r(model)
}
