% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_percentile_to_z.R
\name{convert_z_to_percentile}
\alias{convert_z_to_percentile}
\alias{convert_percentile_to_z}
\alias{z_to_percentile}
\alias{percentile_to_z}
\title{Z score to Percentile}
\usage{
convert_z_to_percentile(z)

convert_percentile_to_z(percentile)

z_to_percentile(z)

percentile_to_z(percentile)
}
\arguments{
\item{z, percentile}{Z score or percentile.}
}
\description{
Convert between Z scores (values expressed in terms of standard deviation) to percentiles (the proportion of a normal distribution below).
}
\examples{
z_to_percentile(1.96)
percentile_to_z(0.975)

}
