% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_between_common_language.R
\name{d_to_cles}
\alias{d_to_cles}
\alias{convert_d_to_common_language,}
\alias{d_to_common_language,}
\alias{rbs_to_common_language,}
\alias{convert_rbs_to_common_language}
\alias{rbs_to_cles}
\title{Convert Standardized Mean Difference to Common Language Effect Sizes}
\usage{
d_to_cles(d)

rbs_to_cles(rbs)
}
\arguments{
\item{d, rbs}{A numeric value of Cohen's d / rank-biserial correlation \emph{or}
the output from \code{\link[=cohens_d]{cohens_d()}} / \code{\link[=rank_biserial]{rank_biserial()}}.}
}
\value{
A list of \verb{Cohen's U3}, \code{Overlap}, \verb{Probability of superiority}, a
numeric vector of \verb{Probability of superiority}, or a data frame, depending
on the input.
}
\description{
Convert Standardized Mean Difference to Common Language Effect Sizes
}
\details{
This function use the following formulae for Cohen's \emph{d}:
\deqn{Cohen's U_3 = \Phi(d)}{U3 = pnorm(d)}
\cr\cr
\deqn{Overlap = 2 \times \Phi(-|d|/2)}{Overlap = 2 * pnorm(-abs(d) / 2)}
\cr\cr
\deqn{Pr(superiority) = \Phi(d/\sqrt{2})}{Pr(superiority) = pnorm(d / sqrt(2))}
}
\note{
These calculations assume that the populations have equal variance and are
normally distributed.
}
\references{
\itemize{
\item Cohen, J. (1977). Statistical power analysis for the behavioral sciences.
New York: Routledge.
\item Reiser, B., & Faraggi, D. (1999). Confidence intervals for the overlapping
coefficient: the normal equal variance case. Journal of the Royal Statistical
Society, 48(3), 413-418.
\item Ruscio, J. (2008). A probability-based measure of effect size: robustness
to base rates and other factors. Psychological methods, 13(1), 19–30.
}
}
\seealso{
\code{\link[=cohens_d]{cohens_d()}}, \code{\link[=rank_biserial]{rank_biserial()}}

Other convert between effect sizes: 
\code{\link{d_to_r}()},
\code{\link{eta2_to_f2}()},
\code{\link{odds_to_probs}()},
\code{\link{oddsratio_to_riskratio}()}
}
\concept{convert between effect sizes}
