% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{desirability}
\alias{desirability}
\title{Desirability of Poop vs Chocolate}
\format{
A data frame with 489 rows and 3 variables:
\describe{
\item{participant_id}{Unique identifier for each participant}
\item{desirability_poop}{Desirability rating of poop (1 = very undesirable, 7 = very desirable)}
\item{desirability_chocolate}{Desirability rating of chocolate (1 = very undesirable, 7 = very desirable)}
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data("desirability")
head(desirability, n = 5)
#>   participant_id poop chocolate
#> 1          088cf    4         6
#> 2          bcae1    1         7
#> 3          d21b0    1         7
#> 4          2019a    1         7
#> 5          c1b73    1         7
}\if{html}{\out{</div>}}
}
\description{
A subset of Ian Hussey's replication of \emph{Balcetis & Dunning (2010) Study 3b}.
Data available at https://github.com/ianhussey/not-so-simple-preferences.
}
\seealso{
Other effect size datasets: 
\code{\link{Music_preferences}},
\code{\link{Music_preferences2}},
\code{\link{RCT_table}},
\code{\link{Smoking_FASD}},
\code{\link{food_class}},
\code{\link{hardlyworking}},
\code{\link{rouder2016}},
\code{\link{screening_test}}
}
\concept{effect size datasets}
\keyword{data}
