% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_API.R, R/eta_squared.R
\name{effectsize_API}
\alias{effectsize_API}
\alias{.es_aov}
\alias{.es_aovlist}
\alias{.es_anova}
\title{\code{effectsize} API}
\usage{
.es_aov(
  aov_table,
  type = c("eta", "omega", "epsilon"),
  partial = TRUE,
  generalized = FALSE,
  ci = 0.95,
  alternative = "greater",
  verbose = TRUE,
  include_intercept = FALSE
)

.es_aovlist(
  aov_table,
  DV_names,
  type = c("eta", "omega", "epsilon"),
  partial = TRUE,
  generalized = FALSE,
  ci = 0.95,
  alternative = "greater",
  verbose = TRUE,
  include_intercept = FALSE
)

.es_anova(
  aov_table,
  type = c("eta", "omega", "epsilon"),
  partial = TRUE,
  generalized = FALSE,
  ci = 0.95,
  alternative = "greater",
  verbose = TRUE,
  include_intercept = FALSE
)
}
\arguments{
\item{type}{Which effect size to compute?}

\item{partial, generalized, ci, alternative, verbose}{See \code{\link[=eta_squared]{eta_squared()}}.}

\item{include_intercept}{Should the intercept (\code{(Intercept)}) be included?}

\item{DV_names}{A character vector with the names of all the predictors,
including the grouping variable (e.g., \code{"Subject"}).}
}
\description{
\code{effectsize} API
}
\section{Adding support for ANOVA effect sizes (\code{eta_squared()} functions)}{


To add support for you model, create a new \code{.anova_es()} method function.
This functions should generally do 3 things:
\enumerate{
\item Build a data frame with all the required information.
\item Pass the data frame to one of the 3 functions.
\item Set some attributes to the output.
}
\subsection{\code{.es_aov()} for simple ANOVA tables}{

The input data frame must have these columns:
\itemize{
\item \code{Parameter} (char) - The name of the parameter or, more often, the term.
\item \code{Sum_Squares} (num) - The sum of squares.
\item \code{df} (num) - The degrees of freedom associated with the \code{Sum_Squares}.
\item \code{Mean_Square_residuals} (num; optional) - if \emph{not} present, is calculated as \code{Sum_Squares / df}.
(Any other column is ignored.)
}

And exactly \emph{1} row Where \code{Parameter} is \code{Residual}.

Optionally, one of the rows can have a \code{(Intercept)} value for \code{Parameter}.

E.g. of a minimally valid data.frame:\if{html}{\out{<div class="sourceCode r">}}\preformatted{min_aov <- data.frame(
  Parameter = c("(Intercept)", "A", "B", "Residuals"),
  Sum_Squares = c(30, 40, 10, 100),
  df = c(1, 1, 2, 50)
)
}\if{html}{\out{</div>}}

Pass the data frame to \code{.es_aov()}:\if{html}{\out{<div class="sourceCode r">}}\preformatted{.es_aov(
  min_aov,
  type = "eta",
  partial = TRUE,
  generalized = FALSE,
  include_intercept = FALSE,
  ci = 0.95, alternative = "greater",
  verbose = TRUE
)
}\if{html}{\out{</div>}}\preformatted{##   Parameter Eta2_partial   CI    CI_low CI_high
## 2         A   0.28571429 0.95 0.1232247       1
## 3         B   0.09090909 0.95 0.0000000       1
}

The output is a data frame with the columns: \code{Parameter}, the effect size,
and (optioanlly) \code{CI} + \code{CI_low} + \code{CI_high},

And with the following attributes: \code{partial}, \code{generalized}, \code{ci},
\code{alternative}, \code{anova_type} (\code{NA} or \code{NULL}), \code{approximate}.

You must then add the \code{anova_type} attribute to {1, 2, 3, or \code{NA}} and return
the output.
}
}

