% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_extra.R
\name{effectsize_options}
\alias{effectsize_options}
\title{\code{effectsize} options}
\description{
Currently, the following global options are supported:
\itemize{
\item \code{es.use_symbols} \link{logical}: Should proper symbols be printed (\code{TRUE}) instead of transliterated effect size names (\code{FALSE}; default).
}
}
