% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_test.R
\name{equivalence_test.effectsize_table}
\alias{equivalence_test.effectsize_table}
\title{Test for Practical Equivalence}
\usage{
\method{equivalence_test}{effectsize_table}(x, range = "default", ...)
}
\arguments{
\item{x}{An effect size table, such as returned by \code{cohens_d, eta_squared, F_to_r}, etc.}

\item{range}{The range of practical equivalence of an effect. If a single value is provided,
the test is done against \code{c(-range, range)}. For effect sizes that cannot be negative,
the lower bound is set to 0. If \code{"default"}, will be set to \code{[-.1, .1]}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame.
}
\description{
Perform a \strong{Test for Practical Equivalence} for indices of effect size.
}
\examples{

\donttest{
model <- aov(mpg ~ factor(am) * factor(cyl), data = mtcars)
es <- eta_squared(model)
equivalence_test(es, range = 0.15)

ds <- t_to_d(t = c(0.45, -0.65, -2.2, 2.25),
             df_error = c(675, 525, 900, 1875))
(equi <- equivalence_test(ds, range = 0.2))

# Can also plot
if (require(see)) plot(equi)
}

}
\references{
\itemize{
  \item Campbell, H., & Gustafson, P. (2018). Conditional equivalence testing: An alternative remedy for publication bias. PLOS ONE, 13(4), e0195145. https://doi.org/10.1371/journal.pone.0195145
  \item Lakens, D. (2017). Equivalence Tests: A Practical Primer for t Tests, Correlations, and Meta-Analyses. Social Psychological and Personality Science, 8(4), 355–362. https://doi.org/10.1177/1948550617697177
}
}
\seealso{
For more details, see \code{\link[bayestestR:equivalence_test]{equivalence_test}}.
}
