% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_effectsize_name.R
\docType{data}
\name{es_info}
\alias{es_info}
\title{List of effect size names}
\format{
An object of class \code{data.frame} with 37 rows and 6 columns.
}
\usage{
es_info
}
\description{
Can always add more info here if need be...
}
\keyword{internal}
