% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.lavaan.R
\name{interpret.lavaan}
\alias{interpret.lavaan}
\title{Interpretation for lavaan objects}
\usage{
\method{interpret}{lavaan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lavaan}.}
}
\description{
Interpretation for lavaan objects
}
\examples{
library(effectsize)

# Structural Equation Models (SEM)
if (require("lavaan")) {
  structure <- " ind60 =~ x1 + x2 + x3
                 dem60 =~ y1 + y2 + y3
                 dem60 ~ ind60 "
  model <- lavaan::sem(structure, data = PoliticalDemocracy)
  interpret(model)
}
}
