% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_direction.R
\name{interpret_direction}
\alias{interpret_direction}
\title{Interpret Direction}
\usage{
interpret_direction(x)
}
\arguments{
\item{x}{Numeric value.}
}
\description{
Interpret Direction
}
\examples{
interpret_direction(.02)
interpret_direction(c(.5, -.02))

}
\keyword{interpreters}
