% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_kendalls_w.R
\name{interpret_kendalls_w}
\alias{interpret_kendalls_w}
\title{Interpret Kendall's Coefficient of Concordance \emph{W}}
\usage{
interpret_kendalls_w(w, rules = "landis1977")
}
\arguments{
\item{w}{Value or vector of Kendall's coefficient of concordance.}

\item{rules}{Can be \code{"landis1977"} (default) or a custom set of \code{\link[=rules]{rules()}}.}
}
\description{
Interpret Kendall's Coefficient of Concordance \emph{W}
}
\section{Rules}{


Landis & Koch (1977) (\code{landis1977}; default):
\itemize{
\item \strong{W < 0.2} - Slight agreement
\item \strong{0.2 <= W < 0.4} - Fair agreement
\item \strong{0.4 <= W < 0.6} - Moderate agreement
\item \strong{0.6 <= W < 0.8} - Substantial agreement
\item \strong{0.8 <= W} - Almost perfect agreement
}
}

\references{
\itemize{
\item Landis, J. R., & Koch G. G. (1977). The measurement of observer agreement
for categorical data. Biometrics, 33:159-74.
}
}
\keyword{interpreters}
