% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_oddsratio.R
\name{interpret_oddsratio}
\alias{interpret_oddsratio}
\title{Interpret Odds Ratio}
\usage{
interpret_oddsratio(OR, rules = "cohen1988", p0 = NULL, log = FALSE, ...)
}
\arguments{
\item{OR}{Value or vector of (log) odds ratio values.}

\item{rules}{If \code{"cohen1988"} (default), \code{OR} is transformed to a
standardized difference (via \code{\link[=oddsratio_to_d]{oddsratio_to_d()}}) and interpreted according
to Cohen's rules (see \code{\link[=interpret_cohens_d]{interpret_cohens_d()}}; see Chen et al., 2010). If a
custom set of \code{\link[=rules]{rules()}} is used, OR is interpreted as is.}

\item{p0}{Baseline risk. If not specified, the \emph{d} to \emph{OR} conversion uses am approximation (see details).}

\item{log}{Are the provided values log odds ratio.}

\item{...}{Currently not used.}
}
\description{
Interpret Odds Ratio
}
\section{Rules}{


Rules apply to OR as ratios, so OR of 10 is as extreme as a OR of 0.1 (1/10).
\itemize{
\item Cohen (1988) (\code{"cohen1988"}, based on the \code{\link[=oddsratio_to_d]{oddsratio_to_d()}} conversion, see \code{\link[=interpret_cohens_d]{interpret_cohens_d()}})
\itemize{
\item \strong{OR < 1.44} - Very small
\item \strong{1.44 <= OR < 2.48} - Small
\item \strong{2.48 <= OR < 4.27} - Medium
\item \strong{OR >= 4.27} - Large
}
}
}

\examples{
interpret_oddsratio(1)
interpret_oddsratio(c(5, 2))

}
\references{
\itemize{
\item Cohen, J. (1988). Statistical power analysis for the behavioral sciences
(2nd Ed.). New York: Routledge.
\item Chen, H., Cohen, P., & Chen, S. (2010). How big is a big odds ratio?
Interpreting the magnitudes of odds ratios in epidemiological studies.
Communications in Statistics-Simulation and Computation, 39(4), 860-864.
\item Sánchez-Meca, J., Marín-Martínez, F., & Chacón-Moscoso, S. (2003).
Effect-size indices for dichotomized outcomes in meta-analysis. Psychological
methods, 8(4), 448.
}
}
\keyword{interpreters}
