% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_p.R
\name{interpret_p}
\alias{interpret_p}
\title{Interpret \emph{p}-Values}
\usage{
interpret_p(p, rules = "default")
}
\arguments{
\item{p}{Value or vector of p-values.}

\item{rules}{Can be \code{"default"}, \code{"rss"} (for \emph{Redefine statistical
significance} rules) or custom set of \code{\link[=rules]{rules()}}.}
}
\description{
Interpret \emph{p}-Values
}
\section{Rules}{


Default (\code{default}):
\itemize{
\item \strong{p < 0.05} - Significant
\item \strong{0.05 <= p} - Not significant
}

RSS (\code{rss}):
\itemize{
\item \strong{p < 0.005} - Significant
\item \strong{0.005 <= p < 0.05} - Suggestive
\item \strong{0.05 <= p} - Not significant
}
}

\examples{
interpret_p(c(.5, .02, 0.001))
interpret_p(c(.5, .02, 0.001), rules = "rss")

}
\references{
\itemize{
\item Benjamin, D. J., Berger, J. O., Johannesson, M., Nosek, B. A., Wagenmakers, E. J., Berk, R., ... & Cesarini, D. (2018). Redefine statistical significance. Nature Human Behaviour, 2(1), 6-10.
}
}
\keyword{interpreters}
