% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_pd.R
\name{interpret_pd}
\alias{interpret_pd}
\title{Interpret Probability of Direction (pd)}
\usage{
interpret_pd(pd, rules = "default", ...)
}
\arguments{
\item{pd}{Value or vector of probabilities of direction.}

\item{rules}{Can be \code{"default"}, \code{"makowski2019"} or a custom set of
\code{\link[=rules]{rules()}}.}

\item{...}{Not directly used.}
}
\description{
Interpret Probability of Direction (pd)
}
\section{Rules}{


Default (i.e., equivalent to p-values):
\itemize{
\item \strong{pd <= 0.975} - Not significant
\item \strong{0.975 < pd} - Significant
}

Makowski et al. (2019) (\code{makowski2019}):
\itemize{
\item \strong{pd <= 0.95} - Uncertain
\item \strong{0.95 < pd <= 0.97} - Possibly existing
\item \strong{0.97 < pd <= 0.99} - Likely existing
\item \strong{0.99 < pd <= 0.999} - Probably existing
\item \strong{0.999 < pd} - Certainly existing
}
}

\examples{
interpret_pd(.98)
interpret_pd(c(.96, .99), rules = "makowski2019")
}
\references{
\itemize{
\item Makowski, D., Ben-Shachar, M. S., Chen, S. H., and Lüdecke, D. (2019). Indices of effect existence and significance in the Bayesian framework. Frontiers in psychology, 10, 2767.
}
}
\keyword{interpreters}
