% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_effectsize_name.R
\name{is_effectsize_name}
\alias{is_effectsize_name}
\title{Checks if character is of a supported effect size}
\usage{
is_effectsize_name(x)
}
\arguments{
\item{x}{A character, or a vactor / list of ones.}
}
\description{
For use by other functions and packages.
}
