% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\name{med}
\alias{med}
\title{Most Extreme Differences}
\usage{
med(
  x,
  y = NULL,
  data = NULL,
  pdf,
  mu = 0,
  reference = NULL,
  ci = 0.95,
  alternative = "two.sided",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x, y}{A numeric vector, or a character name of one in \code{data}.
Any missing values (\code{NA}s) are dropped from the resulting vector.
\code{x} can also be a formula (see \code{\link[stats:t.test]{stats::t.test()}}), in which case \code{y} is
ignored.}

\item{data}{An optional data frame containing the variables.}

\item{pdf}{a character string naming a cumulative distribution function or an
actual cumulative distribution function such as \code{pnorm}. Only continuous
CDFs are valid}

\item{mu}{a number indicating the true value of the mean (or
    difference in means if you are performing a two sample test).}

\item{reference}{(Optional) character value of the "group" used as the
reference. By default, the \emph{second} group is the reference group.}

\item{ci}{Confidence Interval (CI) level}

\item{alternative}{a character string specifying the alternative hypothesis.
This \emph{does not} control the type of CI returned, but the effect calculated;
see \code{\link[stats:ks.test]{stats::ks.test()}} for more details.}

\item{verbose}{Toggle warnings and messages on or off.}

\item{...}{for the default method, parameters of the distribution specified
(as a character string) by \code{pdf}.}
}
\value{
A data frame with the effect size \code{D} and its CI (\code{CI_low} and
\code{CI_high}).
}
\description{
Compute the Most Extreme Differences effect sizes between two ECDFs. Pair
with any reported Kolmogorov-Smirnov test (\code{\link[stats:ks.test]{stats::ks.test()}}).
}
\details{
Something...
}
\section{CIs and Significance Tests}{
"Confidence intervals on measures of effect size convey all the information
in a hypothesis test, and more." (Steiger, 2004). Confidence (compatibility)
intervals and p values are complementary summaries of parameter uncertainty
given the observed data. A dichotomous hypothesis test could be performed
with either a CI or a p value. The 100 (1 - \eqn{\alpha})\% confidence
interval contains all of the parameter values for which \emph{p} > \eqn{\alpha}
for the current data and model. For example, a 95\% confidence interval
contains all of the values for which p > .05.
\cr\cr
Note that a confidence interval including 0 \emph{does not} indicate that the null
(no effect) is true. Rather, it suggests that the observed data together with
the model and its assumptions combined do not provided clear evidence against
a parameter value of 0 (same as with any other value in the interval), with
the level of this evidence defined by the chosen \eqn{\alpha} level (Rafi &
Greenland, 2020; Schweder & Hjort, 2016; Xie & Singh, 2013). To infer no
effect, additional judgments about what parameter values are "close enough"
to 0 to be negligible are needed ("equivalence testing"; Bauer & Kiesser,
1996).
}

\section{Plotting with \code{see}}{

The \code{see} package contains relevant plotting functions. See the \href{https://easystats.github.io/see/articles/effectsize.html}{plotting vignette in the \code{see} package}.
}

\examples{
x <- rnorm(50)
y <- runif(30)

# Do x and y come from the same distribution?
ks.test(x, y)
med(x, y)

# Is the CDS of x larger than that of y?
ks.test(x, y, alternative = "l")
med(x, y, alternative = "l")

plot(ecdf(x), xlim = range(c(x, y)), col = "blue")
lines(ecdf(y), col = "red")

# Does x come from a shifted gamma distribution with shape 3 and rate 2?
ks.test(x+2, "pgamma", 3, 2)
med(x+2, pdf = "pgamma", shape = 3, rate = 2)

plot(ecdf(x + 2), xlim = range(x + 2), col = "blue")
curve(pgamma(x, 3, 2), col = "red", add = TRUE)


}
\seealso{
Other standardized differences: 
\code{\link{cohens_d}()},
\code{\link{mahalanobis_d}()},
\code{\link{means_ratio}()},
\code{\link{p_superiority}()},
\code{\link{rank_biserial}()},
\code{\link{repeated_measures_d}()}
}
\concept{standardized differences}
