% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Music_preferences}
\alias{Music_preferences}
\title{Music Preference by College Major}
\format{
A 4-by-3 table, with a \emph{column} for each major and a \emph{row} for each type of music.
}
\description{
Fictional data.
}
\keyword{data}
