% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Music_preferences2}
\alias{Music_preferences2}
\title{Music Preference by College Major}
\format{
A 4-by-3 table, with a \emph{column} for each major and a \emph{row} for each type of music.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data("Music_preferences2")
Music_preferences2
#>       Pop Rock Jazz Classic
#> Psych 151  130   12       7
#> Econ   77    6  111       4
#> Law     0    4    2     165
}\if{html}{\out{</div>}}
}
\description{
Fictional data, with more extreme preferences than \link{Music_preferences}
}
\seealso{
Other effect size datasets: 
\code{\link{Music_preferences}},
\code{\link{RCT_table}},
\code{\link{Smoking_FASD}},
\code{\link{food_class}},
\code{\link{hardlyworking}},
\code{\link{rouder2016}},
\code{\link{screening_test}}
}
\concept{effect size datasets}
\keyword{data}
