% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_between_riskchange_ro_probs.R
\name{oddsratio_to_odds}
\alias{oddsratio_to_odds}
\alias{oddsratio_to_probs}
\title{Convert Between Metrics of Change in Probabilities and Probabilities}
\usage{
oddsratio_to_odds(OR, p0, Odds0, log = FALSE, ...)

oddsratio_to_probs(OR, p0, Odds0, log = FALSE, ...)
}
\arguments{
\item{OR}{Odds ratio \code{odds(p1)/odds(p0)}}

\item{p0, Odds0}{Base rate or odds. Exactly one must be specified.}

\item{log}{If \code{TRUE}, inputs \code{OR} and \code{Odds0} are taken to be \code{logOR} and
\code{logOdds0}, and outputted odds are \code{logOdds1}.}

\item{...}{Ignored}
}
\description{
Convert Between Metrics of Change in Probabilities and Probabilities
}
