% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{preferences}
\alias{preferences}
\title{Preferences of Poop vs Chocolate}
\format{
A data frame with 489 rows and 3 variables:
\describe{
\item{participant_id}{Unique identifier for each participant}
\item{poop}{Desirability rating of poop (1 = very undesirable, 7 = very desirable)}
\item{chocolate}{Desirability rating of chocolate (1 = very undesirable, 7 = very desirable)}
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data("preferences")
head(preferences, n = 5)
#>   participant_id     poop chocolate
#> 1          088cf 4.000000         6
#> 2          bcae1 2.333333         7
#> 3          d21b0 1.000000         7
#> 4          2019a 1.000000         7
#> 5          c1b73 3.000000         7
}\if{html}{\out{</div>}}
}
\description{
A subset of Hussey and Cummins replication of \emph{Balcetis & Dunning (2010)
Study 3b}. Each scale is average of 3 items both for "a toilet filled with
human poop" and for  and for "a toilet filled with human poop" (how positive
or negative / pleasant or unpleasant / good or bad).
}
\references{
Hussey, I., & Cummins, J. (2025). (Not so) simple preferences. https://github.com/ianhussey/not-so-simple-preferences
}
\seealso{
\code{\link[=repeated_measures_d]{repeated_measures_d()}}

Other effect size datasets: 
\code{\link{Music_preferences}},
\code{\link{Music_preferences2}},
\code{\link{RCT_table}},
\code{\link{Smoking_FASD}},
\code{\link{food_class}},
\code{\link{hardlyworking}},
\code{\link{rouder2016}},
\code{\link{screening_test}}
}
\concept{effect size datasets}
\keyword{data}
