% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/print.effectsize_table.R
\name{plot.effectsize_table}
\alias{plot.effectsize_table}
\alias{print.effectsize_table}
\alias{format.effectsize_table}
\alias{print.effectsize_difference}
\title{Methods for \code{effectsize} tables}
\usage{
\method{plot}{effectsize_table}(x, ...)

\method{print}{effectsize_table}(x, digits = 2, ...)

\method{format}{effectsize_table}(x, digits = 2, ...)

\method{print}{effectsize_difference}(x, digits = 2, append_CL = FALSE, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Arguments passed to or from other functions.}

\item{digits}{Number of significant digits.}

\item{append_CL}{Should the Common Language Effect Sizes be printed as well?
Not applicable to Glass' Delta (See \code{\link[=d_to_common_language]{d_to_common_language()}})}
}
\description{
Printing, formatting and plotting methods for \code{effectsize} tables.
}
