% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RCT_table}
\alias{RCT_table}
\title{Fictional Results from a Workers' Randomized Control Trial}
\format{
A 2-by-2 table, with a \emph{column} for each group and a \emph{row} for the diagnosis.
}
\description{
Fictional Results from a Workers' Randomized Control Trial
}
\keyword{data}
