% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{equivalence_test}
\alias{standardize}
\alias{standardise}
\alias{standardize_parameters}
\alias{standardize_posteriors}
\alias{standardize_info}
\alias{display}
\alias{print_html}
\alias{print_md}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR:equivalence_test]{equivalence_test()}}}

  \item{datawizard}{\code{\link[datawizard:standardise]{standardise()}}, \code{\link[datawizard:standardize]{standardize()}}}

  \item{insight}{\code{\link[insight:display]{display()}}, \code{\link[insight:print_html]{print_html()}}, \code{\link[insight:print_md]{print_md()}}}

  \item{parameters}{\code{\link[parameters:standardize_info]{standardize_info()}}, \code{\link[parameters:standardize_parameters]{standardize_parameters()}}, \code{\link[parameters:standardize_posteriors]{standardize_posteriors()}}}
}}

