% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_d.R
\name{repeated_measures_d}
\alias{repeated_measures_d}
\alias{rm_d}
\title{Standardized Mean Differences for Repeated Measures}
\usage{
repeated_measures_d(
  x,
  y,
  data = NULL,
  mu = 0,
  type = c("rm", "av", "z", "b", "d", "r"),
  adjust = TRUE,
  ci = 0.95,
  alternative = "two.sided",
  verbose = TRUE,
  ...
)

rm_d(
  x,
  y,
  data = NULL,
  mu = 0,
  type = c("rm", "av", "z", "b", "d", "r"),
  adjust = TRUE,
  ci = 0.95,
  alternative = "two.sided",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x, y}{Paired numeric vectors, or names of ones in \code{data}. \code{x} can also
be a formula:
\itemize{
\item \code{Pair(x,y) ~ 1} for wide data.
\item \code{y ~ condition | id} for long data, possibly with repetitions.
}}

\item{data}{An optional data frame containing the variables.}

\item{mu}{a number indicating the true value of the mean (or
    difference in means if you are performing a two sample test).}

\item{type}{Type of repeated measures standardized differences. See details.}

\item{adjust}{Apply Hedges' small-sample bias correction?}

\item{ci}{Confidence Interval (CI) level}

\item{alternative}{a character string specifying the alternative hypothesis;
Controls the type of CI returned: \code{"two.sided"} (default, two-sided CI),
\code{"greater"} or \code{"less"} (one-sided CI). Partial matching is allowed (e.g.,
\code{"g"}, \code{"l"}, \code{"two"}...). See \emph{One-Sided CIs} in \link{effectsize_CIs}.}

\item{verbose}{Toggle warnings and messages on or off.}

\item{...}{Arguments passed to or from other methods. When \code{x} is a formula,
these can be \code{subset} and \code{na.action}.}
}
\value{
A data frame with the effect size and their CIs (\code{CI_low} and
\code{CI_high}).
}
\description{
Short description... TODO. Pair with any reported \code{stats::t.test(paired = TRUE)}.
}
\note{
\code{rm_d()} is an alias for \code{repeated_measures_d()}.
}
\section{Types of TODO}{
TODO
}

\section{Confidence (Compatibility) Intervals (CIs)}{
Confidence intervals are estimated using the standard normal parametric
method (see TODO).
}

\section{CIs and Significance Tests}{
"Confidence intervals on measures of effect size convey all the information
in a hypothesis test, and more." (Steiger, 2004). Confidence (compatibility)
intervals and p values are complementary summaries of parameter uncertainty
given the observed data. A dichotomous hypothesis test could be performed
with either a CI or a p value. The 100 (1 - \eqn{\alpha})\% confidence
interval contains all of the parameter values for which \emph{p} > \eqn{\alpha}
for the current data and model. For example, a 95\% confidence interval
contains all of the values for which p > .05.
\cr\cr
Note that a confidence interval including 0 \emph{does not} indicate that the null
(no effect) is true. Rather, it suggests that the observed data together with
the model and its assumptions combined do not provided clear evidence against
a parameter value of 0 (same as with any other value in the interval), with
the level of this evidence defined by the chosen \eqn{\alpha} level (Rafi &
Greenland, 2020; Schweder & Hjort, 2016; Xie & Singh, 2013). To infer no
effect, additional judgments about what parameter values are "close enough"
to 0 to be negligible are needed ("equivalence testing"; Bauer & Kiesser,
1996).
}

\section{Plotting with \code{see}}{

The \code{see} package contains relevant plotting functions. See the \href{https://easystats.github.io/see/articles/effectsize.html}{plotting vignette in the \code{see} package}.
}

\examples{
# Paired data -------

data("sleep")
sleep2 <- reshape(sleep, direction = "wide",
                  idvar = "ID", timevar = "group")

repeated_measures_d(Pair(extra.1, extra.2) ~ 1, data = sleep2)

# Same as:
# repeated_measures_d(sleep$extra[sleep$group==1],
#                     sleep$extra[sleep$group==2])
# repeated_measures_d(extra ~ group | ID, data = sleep)


# More options:
repeated_measures_d(Pair(extra.1, extra.2) ~ 1, data = sleep2, mu = -1)
repeated_measures_d(Pair(extra.1, extra.2) ~ 1, data = sleep2, alternative = "less")

# Other types
repeated_measures_d(Pair(extra.1, extra.2) ~ 1, data = sleep2, type = "av")
repeated_measures_d(Pair(extra.1, extra.2) ~ 1, data = sleep2, type = "b")
repeated_measures_d(Pair(extra.1, extra.2) ~ 1, data = sleep2, type = "d")
repeated_measures_d(Pair(extra.1, extra.2) ~ 1, data = sleep2, type = "z", adjust = FALSE)

# d_z is the same as Cohen's d for one sample (of individual difference):
cohens_d(extra.1 - extra.2 ~ 1, data = sleep2)



# Repetition data -----------

data("rouder2016")

# For types rm, ad, z, b, data is aggragated
repeated_measures_d(rt ~ cond | id, data = rouder2016)

# same as:
rouder2016_wide <- tapply(rouder2016[["rt"]], rouder2016[1:2], mean)
repeated_measures_d(rouder2016_wide[, 1], rouder2016_wide[, 2])

# For types d and r, data is not aggragated:
repeated_measures_d(rt ~ cond | id, data = rouder2016, type = "d")
repeated_measures_d(rt ~ cond | id, data = rouder2016, type = "r")



}
\references{
TODO
}
\seealso{
\code{\link[=cohens_d]{cohens_d()}}

Other standardized differences: 
\code{\link{cohens_d}()},
\code{\link{mahalanobis_d}()},
\code{\link{means_ratio}()},
\code{\link{p_superiority}()},
\code{\link{rank_biserial}()}
}
\concept{standardized differences}
