% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rouder2016}
\alias{rouder2016}
\title{Jeff Rouder's Example Dataset for Repeated Measures}
\format{
A data frame with 2500 rows and 3 variables:
\describe{
\item{id}{participant: 1...25}
\item{cond}{condition: 1,2}
\item{rt}{response time in seconds}
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data("rouder2016")
head(rouder2016, n = 5)
#>   id cond    rt
#> 1  1    1 0.560
#> 2  1    1 0.930
#> 3  1    1 0.795
#> 4  1    1 0.615
#> 5  1    1 1.028
}\if{html}{\out{</div>}}
}
\description{
A dataset "with 25 people each observing 50 trials in 2 conditions",
published as \code{effectSizePuzzler.txt} by Jeff Rouder on March 24, 2016
(\emph{http://jeffrouder.blogspot.com/2016/03/the-effect-size-puzzler.html}).
\cr\cr
The data is used in examples and tests of \code{\link[=rm_d]{rm_d()}}.
}
\seealso{
Other effect size datasets: 
\code{\link{Music_preferences}},
\code{\link{Music_preferences2}},
\code{\link{RCT_table}},
\code{\link{Smoking_FASD}},
\code{\link{desirability}},
\code{\link{food_class}},
\code{\link{hardlyworking}},
\code{\link{screening_test}}
}
\concept{effect size datasets}
\keyword{data}
