% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Smoking_FASD}
\alias{Smoking_FASD}
\title{Frequency of FASD for Smoking Mothers}
\format{
A 1-by-3 table, with a \emph{column} for each diagnosis.
}
\description{
Fictional data.
}
\keyword{data}
