% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Test_results}
\alias{Test_results}
\title{Results from 2 Screening Tests}
\format{
A data frame with 1600 rows and 3 variables:
\describe{
\item{Diagnosis}{Ground truth}
\item{Test1}{Results given by the 1st test}
\item{Test2}{Results given by the 2nd test}
}
}
\description{
A sample (simulated) dataset, used in tests and some examples.
}
\keyword{data}
