% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_ratio.R
\name{variance_ratio}
\alias{variance_ratio}
\title{Ratio of Variances}
\usage{
variance_ratio(
  x,
  y = NULL,
  data = NULL,
  paired = FALSE,
  log = FALSE,
  ci = 0.95,
  alternative = "two.sided",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x, y}{A numeric vector, or a character name of one in \code{data}.
Any missing values (\code{NA}s) are dropped from the resulting vector.
\code{x} can also be a formula (see \code{\link[stats:t.test]{stats::t.test()}}), in which case \code{y} is
ignored.}

\item{data}{An optional data frame containing the variables.}

\item{paired}{If \code{TRUE}, the values of \code{x} and \code{y} are considered as paired.
This produces an effect size that is equivalent to the one-sample effect
size on \code{x - y}. See also \code{\link[=repeated_measures_d]{repeated_measures_d()}} for more options.}

\item{log}{Should the log-ratio be returned? Defaults to \code{FALSE}.
Normally distributed and useful for meta-analysis.}

\item{ci}{Confidence Interval (CI) level}

\item{alternative}{a character string specifying the alternative hypothesis;
Controls the type of CI returned: \code{"two.sided"} (default, two-sided CI),
\code{"greater"} or \code{"less"} (one-sided CI). Partial matching is allowed (e.g.,
\code{"g"}, \code{"l"}, \code{"two"}...). See \emph{One-Sided CIs} in \link{effectsize_CIs}.}

\item{verbose}{Toggle warnings and messages on or off.}

\item{...}{Arguments passed to or from other methods. When \code{x} is a formula,
these can be \code{subset} and \code{na.action}.}
}
\value{
A data frame with the effect size (\code{Vars_ratio}) and its CIs
(\code{CI_low} and \code{CI_high}).
}
\description{
Computes the ratio of two variances of independent or paired samples. For
paired data, this can also be thought of as the ratio of marginal (squared)
scales of a bivariate normal distribution. For independent samples, this is a
convenient wrapper around \code{\link[stats:var.test]{stats::var.test()}} (for a paired version of this
test - the Pitman-Morgan test, see \code{PairedData::Var.test()}).
}
\section{Confidence (Compatibility) Intervals (CIs)}{
For independent samples, confidence intervals are estimated using
\code{\link[stats:var.test]{stats::var.test()}}. For paired samples, confidence intervals are estimated
using the standard parametric method (see Pitman, 1939; Morgan, 1929).
}

\section{CIs and Significance Tests}{
"Confidence intervals on measures of effect size convey all the information
in a hypothesis test, and more." (Steiger, 2004). Confidence (compatibility)
intervals and p values are complementary summaries of parameter uncertainty
given the observed data. A dichotomous hypothesis test could be performed
with either a CI or a p value. The 100 (1 - \eqn{\alpha})\% confidence
interval contains all of the parameter values for which \emph{p} > \eqn{\alpha}
for the current data and model. For example, a 95\% confidence interval
contains all of the values for which p > .05.
\cr\cr
Note that a confidence interval including 0 \emph{does not} indicate that the null
(no effect) is true. Rather, it suggests that the observed data together with
the model and its assumptions combined do not provided clear evidence against
a parameter value of 0 (same as with any other value in the interval), with
the level of this evidence defined by the chosen \eqn{\alpha} level (Rafi &
Greenland, 2020; Schweder & Hjort, 2016; Xie & Singh, 2013). To infer no
effect, additional judgments about what parameter values are "close enough"
to 0 to be negligible are needed ("equivalence testing"; Bauer & Kiesser,
1996).
}

\section{Plotting with \code{see}}{

The \code{see} package contains relevant plotting functions. See the \href{https://easystats.github.io/see/articles/effectsize.html}{plotting vignette in the \code{see} package}.
}

\examples{
# Two Independent Samples ----------

x <- c(0.878, 0.647, 0.598, 2.05, 1.06, 1.29, 1.06, 3.14, 1.29)
y <- c(1.83, 0.50, 1.62, 2.48, 1.68, 1.88, 1.55, 3.06, 1.30)

variance_ratio(x, y)

# More options:
variance_ratio(x, y, alternative = "less")
variance_ratio(x, y, log = TRUE)

# The ratio is scale invariant
variance_ratio(x * 3, y * 3)



# Paired Samples ----------

sleep2 <- reshape(sleep,
  direction = "wide",
  idvar = "ID", timevar = "group"
)
variance_ratio(Pair(extra.1, extra.2) ~ 1, data = sleep2)

}
\references{
\itemize{
\item Morgan, W. A. (1939). A test for the significance of the difference between
the two variances in a sample from a normal bivariate population. Biometrika,
31(1/2), 13-19.
\item Pitman, E. J. G. (1939). A note on normal correlation. Biometrika, 31(1/2), 9-12.
}
}
\seealso{
\code{\link[=means_ratio]{means_ratio()}}, \code{\link[stats:var.test]{stats::var.test()}} and \code{PairedData::Var.test()}
for the \emph{Pitman-Morgan Test} for paired samples.
}
