% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_stat_chisq.R
\name{chisq_to_phi}
\alias{chisq_to_phi}
\alias{chisq_to_cohens_w}
\alias{chisq_to_cramers_v}
\alias{phi_to_chisq}
\title{Conversion Chi-Squared to Phi or Cramer's V}
\usage{
chisq_to_phi(
  chisq,
  n,
  nrow,
  ncol,
  ci = 0.95,
  alternative = "greater",
  adjust = FALSE,
  ...
)

chisq_to_cohens_w(
  chisq,
  n,
  nrow,
  ncol,
  ci = 0.95,
  alternative = "greater",
  adjust = FALSE,
  ...
)

chisq_to_cramers_v(
  chisq,
  n,
  nrow,
  ncol,
  ci = 0.95,
  alternative = "greater",
  adjust = FALSE,
  ...
)

phi_to_chisq(phi, n, ...)
}
\arguments{
\item{chisq}{The Chi-squared statistic.}

\item{n}{Total sample size.}

\item{nrow, ncol}{The number of rows/columns in the contingency table (ignored
for Phi when \code{adjust=FALSE} and \code{CI=NULL}).}

\item{ci}{Confidence Interval (CI) level}

\item{alternative}{a character string specifying the alternative hypothesis;
Controls the type of CI returned: \code{"greater"} (default) or \code{"less"}
(one-sided CI), or \code{"two.sided"} (default, two-sided CI). Partial matching
is allowed (e.g., \code{"g"}, \code{"l"}, \code{"two"}...). See \emph{One-Sided CIs} in
\link{effectsize-CIs}.}

\item{adjust}{Should the effect size be bias-corrected? Defaults to \code{FALSE}.}

\item{...}{Arguments passed to or from other methods.}

\item{phi}{The Phi statistic.}
}
\value{
A data frame with the effect size(s) between 0-1, and confidence
interval(s). See \code{\link[=cramers_v]{cramers_v()}}.
}
\description{
Convert between Chi square (\eqn{\chi^2}), Cramer's V, phi (\eqn{\phi}) and
Cohen's \emph{w} for contingency tables or goodness of fit.
}
\details{
These functions use the following formulae:
\cr
\deqn{\phi = \sqrt{\chi^2 / n}}{phi = sqrt(\chi^2 / n)}
\cr
\deqn{Cramer's V = \phi / \sqrt{min(nrow,ncol)-1}}{Cramer's V = \phi / sqrt(min(nrow,ncol)-1)}
\cr
For adjusted versions, see Bergsma, 2013.
}
\note{
Cohen's \emph{w} is equivalent to \emph{Phi}.
}
\section{Confidence (Compatibility) Intervals (CIs)}{

Unless stated otherwise, confidence (compatibility) intervals (CIs) are
estimated using the noncentrality parameter method (also called the
"pivot method"). This method finds the noncentrality parameter ("\emph{ncp}") of
a noncentral \emph{t}, \emph{F}, or chi-squared distribution that places the observed
\emph{t}, \emph{F}, or chi-squared test statistic at the desired probability point of
the distribution. For example, if the observed \emph{t} statistic is 2.0, with 50
degrees of freedom, for which cumulative noncentral \emph{t} distribution is
\emph{t} = 2.0 the .025 quantile (answer: the noncentral \emph{t} distribution with
\emph{ncp} = .04)? After estimating these confidence bounds on the \emph{ncp}, they are
converted into the effect size metric to obtain a confidence interval for the
effect size (Steiger, 2004).
\cr\cr
For additional details on estimation and troubleshooting, see \link{effectsize_CIs}.
}

\section{CIs and Significance Tests}{

"Confidence intervals on measures of effect size convey all the information
in a hypothesis test, and more." (Steiger, 2004). Confidence (compatibility)
intervals and p values are complementary summaries of parameter uncertainty
given the observed data. A dichotomous hypothesis test could be performed
with either a CI or a p value. The
100(\ifelse{latex}{\out{$1 - \alpha$}}{\ifelse{html}{\out{1 &minus; &alpha;}}{1 - alpha}})\%
confidence interval contains all of the parameter values for which
\ifelse{latex}{\out{$p > \alpha$}}{\ifelse{html}{\out{p > &alpha;}}{p > alpha}}
for the current data and model. For example, a 95\% confidence interval
contains all of the values for which p > .05.
\cr\cr
Note that a confidence interval including 0 \emph{does not} indicate no effect.
Rather, it suggests that the observed data and background model assumptions
combined do not clearly indicate against a parameter value of 0 (or any other
value in the interval), with the level of this evidence defined by the chosen
\ifelse{latex}{\out{$\alpha$}}{\ifelse{html}{\out{&alpha;}}{alpha}} level
(Rafi & Greenland, 2020; Schweder & Hjort, 2016; Xie & Singh, 2013). To
infer no effect, additional judgments about what parameter values are "close
enough" to 0 to be negligible are needed ("equivalence testing";
Bauer & Kiesser, 1996).
}

\section{One-Sided CIs}{

Typically, CIs are constructed as two-tailed intervals, with an equal
proportion of the cumulative probability distribution above and below the
interval. CIs can also be constructed as \emph{one-sided} intervals,
giving only a lower bound or upper bound. This is analogous to computing a
1-tailed \emph{p} value or conducting a 1-tailed hypothesis test.
\cr\cr
Significance tests conducted using CIs (whether a value is inside the interval)
and using \emph{p} values (whether p < alpha for that value) are only guaranteed
to agree when both are constructed using the same number of sides/tails.
\cr\cr
Most effect sizes are not bounded by zero (e.g., \emph{r}, \emph{d}, \emph{g}). These
typically involve \emph{t}- or \emph{z}-statistics and are generally tested using
2-tailed tests and 2-sided CIs.
\cr\cr
Some effect sizes are strictly positive--they have a minimum value of 0.
For example,
\ifelse{latex}{\out{$R^2$}}{\ifelse{html}{\out{*R*<sup>2</sup>}}{R^2}},
\ifelse{latex}{\out{$\eta^2$}}{\ifelse{html}{\out{&eta;<sup>2</sup>}}{eta^2}},
and other variance-accounted-for effect sizes, as well as Cramer's \emph{V} and
multiple \emph{R}, range from 0 to 1. These typically involve \emph{F}- or
\ifelse{latex}{\out{$\chi^2$}}{\ifelse{html}{\out{\chi;<sup>2</sup>}}{chi-squared}}-statistics
and are generally tested using \emph{1-tailed} tests. These test test whether the
estimated effect size is \emph{larger} than the hypothesized value (e.g., 0). The
corresponding CI that yields the same significance decision is a \emph{1-sided} CI
estimating only a lower bound. This is the default CI computed by \emph{effectsize}
for these effect sizes, called by setting \code{alternative = "greater"}.
\cr\cr
This lower bound interval indicates the smallest effect size that is not
significantly different from the observed effect size. That is, it is the
minimum effectsize compatible with the observed data, background model
assumptions, and \ifelse{latex}{\out{$\alpha$}}{\ifelse{html}{\out{&alpha;}}{alpha}}
level. The interval does not indicate a maximum effect size value; anything
up to the maximum possible value of the effect size (e.g., 1) is in the interval.
\cr\cr
An alternative 1-sided CI that can be used to test against the maximum effect
size value (e.g., is
\ifelse{latex}{\out{$R^2$}}{\ifelse{html}{\out{*R*<sup>2</sup>}}{R^2}}
significantly different from a perfect correlation of 1.0?) can by setting
\code{alternative = "less"}. This estimates a CI with only an \emph{upper} bound;
anything from the minimum possible value of the effect size (e.g., 0) up to
this upper bound is in the interval.
\cr\cr
To obtain a 2-sided interval with equal probability proportions above and
below the interval, set \code{alternative = "two-sided"}. These intervals can
be interpreted in the same way as other 2-sided intervals, such as those
for \emph{r}, \emph{d}, or \emph{g}.
\cr\cr
An alternative approach to aligning significance tests using CIs and 1-tailed
\emph{p} values that can often be found in the literature is to
construct a 2-sided CI at a lower confidence level (
100(\ifelse{latex}{\out{$1 - 2\alpha$}}{\ifelse{html}{\out{1 &minus; 2&alpha;}}{1 - 2*alpha}})\%
= \ifelse{latex}{\out{$100 - 2 \times 5\% = 90\%$}}{\ifelse{html}{\out{100 − 2 \%times; 5\% = 90\%}}{100 - 2*5\% = 90\%}}),
estimates the lower bound and upper bound for the above 1-sided intervals
simultaneously. These intervals are commonly reported when conducting equivalence
tests. For example, a 90\% 2-sided interval gives the bounds for an equivalence
test with \ifelse{latex}{\out{$\alpha = .05$}}{\ifelse{html}{\out{&alpha; = .05}}{alpha = .05}}.
However, be aware that this interval does not give 95\% coverage for the
underlying effect size parameter value. For that, construct a 95\% 2-sided CI.
For example:\if{html}{\out{<div class="sourceCode r">}}\preformatted{data("hardlyworking")
fit <- lm(salary ~ n_comps + age, data = hardlyworking)
eta_squared(fit, ci = 0.95, alternative = "less") # default, lower 1-sided bound
}\if{html}{\out{</div>}}\preformatted{## # Effect Size for ANOVA (Type I)
## 
## Parameter | Eta2 (partial) |       95\% CI
## -----------------------------------------
## n_comps   |           0.21 | [0.00, 0.26]
## age       |           0.10 | [0.00, 0.14]
## 
## - One-sided CIs: lower bound fixed at (0).
}\if{html}{\out{<div class="sourceCode r">}}\preformatted{eta_squared(fit, ci = 0.95, alternative = "greater") # upper 1-sided bound
}\if{html}{\out{</div>}}\preformatted{## # Effect Size for ANOVA (Type I)
## 
## Parameter | Eta2 (partial) |       95\% CI
## -----------------------------------------
## n_comps   |           0.21 | [0.16, 1.00]
## age       |           0.10 | [0.06, 1.00]
## 
## - One-sided CIs: upper bound fixed at (1).
}\if{html}{\out{<div class="sourceCode r">}}\preformatted{eta_squared(fit, ci = 0.9, alternative = "two.sided") # both 1-sided bounds for alpha = .05
}\if{html}{\out{</div>}}\preformatted{## # Effect Size for ANOVA (Type I)
## 
## Parameter | Eta2 (partial) |       90\% CI
## -----------------------------------------
## n_comps   |           0.21 | [0.16, 0.26]
## age       |           0.10 | [0.06, 0.14]
}\if{html}{\out{<div class="sourceCode r">}}\preformatted{eta_squared(fit, ci = 0.95, alternative = "two.sided") # 2-sided bounds for alpha = .05
}\if{html}{\out{</div>}}\preformatted{## # Effect Size for ANOVA (Type I)
## 
## Parameter | Eta2 (partial) |       95\% CI
## -----------------------------------------
## n_comps   |           0.21 | [0.15, 0.27]
## age       |           0.10 | [0.06, 0.15]
}
}

\examples{
contingency_table <- as.table(rbind(c(762, 327, 468), c(484, 239, 477), c(484, 239, 477)))

chisq.test(contingency_table)
#
#         Pearson's Chi-squared test
#
# data:  contingency_table
# X-squared = 41.234, df = 4, p-value = 2.405e-08

chisq_to_phi(41.234,
  n = sum(contingency_table),
  nrow = nrow(contingency_table),
  ncol = ncol(contingency_table)
)
chisq_to_cramers_v(41.234,
  n = sum(contingency_table),
  nrow = nrow(contingency_table),
  ncol = ncol(contingency_table)
)
}
\references{
\itemize{
\item Cumming, G., & Finch, S. (2001). A primer on the understanding, use, and
calculation of confidence intervals that are based on central and noncentral
distributions. Educational and Psychological Measurement, 61(4), 532-574.
\item Bergsma, W. (2013). A bias-correction for Cramer's V and Tschuprow's T.
Journal of the Korean Statistical Society, 42(3), 323-328.
}
}
\seealso{
Other effect size from test statistic: 
\code{\link{F_to_eta2}()},
\code{\link{t_to_d}()}
}
\concept{effect size from test statistic}
