% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_language.R
\name{cles}
\alias{cles}
\alias{common_language}
\title{Estimate Common Language Effect Sizes}
\usage{
cles(
  x,
  y = NULL,
  data = NULL,
  mu = 0,
  ci = 0.95,
  alternative = "two.sided",
  verbose = TRUE,
  rank = FALSE,
  ...
)

common_language(
  x,
  y = NULL,
  data = NULL,
  mu = 0,
  ci = 0.95,
  alternative = "two.sided",
  verbose = TRUE,
  rank = FALSE,
  ...
)
}
\arguments{
\item{x}{A formula, a numeric vector, or a character name of one in \code{data}.}

\item{y}{A numeric vector, a grouping (character / factor) vector, a or a
character  name of one in \code{data}. Ignored if \code{x} is a formula.}

\item{data}{An optional data frame containing the variables.}

\item{mu}{a number indicating the true value of the mean (or
    difference in means if you are performing a two sample test).}

\item{ci}{Confidence Interval (CI) level}

\item{alternative}{a character string specifying the alternative hypothesis;
Controls the type of CI returned: \code{"two.sided"} (default, two-sided CI),
\code{"greater"} or \code{"less"} (one-sided CI). Partial matching is allowed (e.g.,
\code{"g"}, \code{"l"}, \code{"two"}...). See \emph{One-Sided CIs} in \link{effectsize_CIs}.}

\item{verbose}{Toggle warnings and messages on or off.}

\item{rank}{Use non-parametric rank-based estimation (see \code{\link[=rank_biserial]{rank_biserial()}})
instead of parametric estimation (see \code{\link[=cohens_d]{cohens_d()}}).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame containing the common language effect sizes (and
optionally their CIs).
}
\description{
Estimate Common Language Effect Sizes
}
\seealso{
\code{\link[=d_to_cles]{d_to_cles()}} \code{\link[=sd_pooled]{sd_pooled()}}

Other effect size indices: 
\code{\link{cohens_d}()},
\code{\link{effectsize.BFBayesFactor}()},
\code{\link{eta_squared}()},
\code{\link{phi}()},
\code{\link{rank_biserial}()},
\code{\link{standardize_parameters}()}
}
\concept{effect size indices}
