% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_language.R
\name{cles}
\alias{cles}
\alias{common_language}
\title{Estimate Common Language Effect Sizes (CLES)}
\usage{
cles(
  x,
  y = NULL,
  data = NULL,
  mu = 0,
  ci = 0.95,
  alternative = "two.sided",
  verbose = TRUE,
  rank = FALSE,
  ...
)

common_language(
  x,
  y = NULL,
  data = NULL,
  mu = 0,
  ci = 0.95,
  alternative = "two.sided",
  verbose = TRUE,
  rank = FALSE,
  ...
)
}
\arguments{
\item{x}{A formula, a numeric vector, or a character name of one in \code{data}.}

\item{y}{A numeric vector, a grouping (character / factor) vector, a or a
character  name of one in \code{data}. Ignored if \code{x} is a formula.}

\item{data}{An optional data frame containing the variables.}

\item{mu}{a number indicating the true value of the mean (or
    difference in means if you are performing a two sample test).}

\item{ci}{Confidence Interval (CI) level}

\item{alternative}{a character string specifying the alternative hypothesis;
Controls the type of CI returned: \code{"two.sided"} (default, two-sided CI),
\code{"greater"} or \code{"less"} (one-sided CI). Partial matching is allowed (e.g.,
\code{"g"}, \code{"l"}, \code{"two"}...). See \emph{One-Sided CIs} in \link{effectsize_CIs}.}

\item{verbose}{Toggle warnings and messages on or off.}

\item{rank}{Use non-parametric rank-based estimation (see \code{\link[=rank_biserial]{rank_biserial()}})
instead of parametric estimation (see \code{\link[=cohens_d]{cohens_d()}}).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame containing the common language effect sizes (and
optionally their CIs).
}
\description{
Estimate Common Language Effect Sizes (CLES)
}
\details{
These measures of effect size present group differences in probabilistic
terms:
\itemize{
\item \strong{Probability of superiority} is the probability that, when sampling an
observation from each of the groups at random, that the observation from
the second group will be larger than the sample from the first group.
\item \strong{Cohen's U3} is the proportion of the second group that is smaller than
the median of the first group.
\item \strong{Overlap} (OVL) is the proportion overlap between the two distributions.
}

For unequal group sizes, it is recommended to use the rank based CLES (\code{rank = TRUE}).
}
\section{Confidence Intervals (CIs)}{

For parametric (\code{rank = FALSE}) CLES, the CIs are transformed CIs for Cohen's
\emph{d} (\code{\link[=d_to_cles]{d_to_cles()}}). For non-parametric (\code{rank = TRUE}) CLES, the CI of
\emph{Pr(superiority)} is a transformed CI of the rank-biserial correlation
(\code{\link[=rb_to_cles]{rb_to_cles()}}), while for Cohen's \emph{U3} and the Overlap coefficient the
confidence intervals are based on the proportion's confidence intervals
returned by \code{\link[stats:prop.test]{stats::prop.test()}}, which give a good close approximation.
}

\examples{
cles(mpg ~ am, data = mtcars)

cles(mpg ~ am, data = mtcars, rank = TRUE)

## Individual CLES
A <- c(0.80, 0.83, 1.89, 1.04, 1.45, 1.38, 1.91, 1.64, 0.73, 1.46)
B <- c(1.15, 0.88, 0.90, 0.74, 1.21)

p_superiority(A, B)

cohens_u3(A, B, rank = TRUE)

p_overlap(A, B)

}
\references{
\itemize{
\item Cohen, J. (1977). Statistical power analysis for the behavioral sciences.
New York: Routledge.
\item Reiser, B., & Faraggi, D. (1999). Confidence intervals for the overlapping
coefficient: the normal equal variance case. Journal of the Royal Statistical
Society, 48(3), 413-418.
\item Ruscio, J. (2008). A probability-based measure of effect size: robustness
to base rates and other factors. Psychological methods, 13(1), 19–30.
}
}
\seealso{
\code{\link[=d_to_cles]{d_to_cles()}} \code{\link[=sd_pooled]{sd_pooled()}}

Other effect size indices: 
\code{\link{cohens_d}()},
\code{\link{effectsize.BFBayesFactor}()},
\code{\link{eta_squared}()},
\code{\link{phi}()},
\code{\link{rank_biserial}()},
\code{\link{standardize_parameters}()}
}
\concept{effect size indices}
