% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d.R
\name{cohens_d}
\alias{cohens_d}
\title{Indices of effect size (Cohen's d, Hedges' g, ...)}
\usage{
cohens_d(x, y = NULL, data = NULL)
}
\arguments{
\item{x}{A continuous variable or a formula.}

\item{y}{A continuous variable, a factor with two groups or a formula.}

\item{data}{An optional data frame (or similar: see model.frame) containing the variables.}
}
\description{
Compute different indices of effect size.
}
\examples{
cohens_d(iris$Sepal.Length, iris$Sepal.Width)

}
