% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_d_to_r.R
\name{d_to_r}
\alias{d_to_r}
\alias{r_to_d}
\alias{convert_d_to_r}
\alias{convert_r_to_d}
\alias{odds_to_d}
\alias{convert_odds_to_d}
\alias{logodds_to_d}
\alias{d_to_odds}
\alias{convert_d_to_odds}
\alias{odds_to_r}
\alias{convert_odds_to_r}
\alias{logodds_to_r}
\alias{r_to_odds}
\alias{convert_r_to_odds}
\title{Convert between \emph{d}, \emph{r} and \emph{Odds ratio}}
\usage{
d_to_r(d, ...)

r_to_d(r, ...)

convert_d_to_r(d, ...)

convert_r_to_d(r, ...)

odds_to_d(odds, log = FALSE, ...)

convert_odds_to_d(odds, log = FALSE, ...)

logodds_to_d(odds, log = TRUE, ...)

d_to_odds(d, log = FALSE, ...)

convert_d_to_odds(d, log = FALSE, ...)

odds_to_r(odds, log = FALSE, ...)

convert_odds_to_r(odds, log = FALSE, ...)

logodds_to_r(odds, log = TRUE, ...)

r_to_odds(r, log = FALSE, ...)

convert_r_to_odds(r, log = FALSE, ...)
}
\arguments{
\item{d}{Standardized difference value (Cohen's d).}

\item{...}{Arguments passed to or from other methods.}

\item{r}{Correlation coefficient r.}

\item{odds}{\emph{Odds ratio} values in vector or data frame.}

\item{log}{Take in or output log odds ratio (such as in logistic models).}
}
\value{
Converted index.
}
\description{
Enables a conversion between different indices of effect size, such as
standardized difference (Cohen's d), correlation r or (log) odds ratios.
}
\details{
\itemize{
\item \emph{d to r}: \eqn{d = \frac{2 * r}{\sqrt{1 - r^2}}}
\item \emph{r to d}: \eqn{r = \frac{d}{\sqrt{d^2 + 4}}}
\item \emph{OR to d}: \eqn{d = \frac{\log(OR)\times\sqrt{3}}{\pi}}
\item \emph{d to OR}: \eqn{log(OR) = d * \frac{\pi}{\sqrt(3)}}
}

Conversions between \emph{OR} and \emph{r} is done through these formulae.
\cr\cr
When converting \emph{d} to \emph{r}, the resulting \emph{r} is also called the binomial
effect size display (BESD; Rosenthal et al., 1982).
}
\examples{
r_to_d(0.5)
d_to_odds(d = 1.154701)
odds_to_r(odds = 8.120534)

d_to_r(d = 1)
r_to_odds(0.4472136, log = TRUE)
odds_to_d(1.813799, log = TRUE)
}
\references{
\itemize{
\item Sánchez-Meca, J., Marín-Martínez, F., & Chacón-Moscoso, S. (2003). Effect-size indices for dichotomized outcomes in meta-analysis. Psychological methods, 8(4), 448.
\item Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009). Converting among effect sizes. Introduction to meta-analysis, 45-49.
\item Rosenthal, R., & Rubin, D. B. (1982). A simple, general purpose display of magnitude of experimental effect. Journal of educational psychology, 74(2), 166.
}
}
