% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_extra.R
\name{effectsize-CIs}
\alias{effectsize-CIs}
\title{Confidence Intervals}
\description{
More information regarding Confidence Intervals and how they are computed in
\code{effectsize}.
}
\section{Confidence Intervals}{

Unless stated otherwise, confidence intervals are estimated using the
Noncentrality parameter method; These methods searches for a the best
non-central parameters (\code{ncp}s) of the noncentral t-, F- or Chi-squared
distribution for the desired tail-probabilities, and then convert these
\code{ncp}s to the corresponding effect sizes. (See full \link{effectsize-CIs} for
more.)
}

\section{CI Contains Zero}{

Keep in mind that \code{ncp} confidence intervals are inverted significance tests,
and only inform us about which values are not significantly different than
our sample estimate. (They do \emph{not} inform us about which values are
plausible, likely or compatible with our data.) Thus, when CIs contain the
value 0, this should \emph{not} be taken to mean that a null effect size is
supported by the data; Instead this merely reflects a non-significant test
statistic - i.e. the \emph{p}-value is greater than alpha (Morey et al., 2016).
\cr\cr
For positive only effect sizes (Eta squared, Cramer's V, etc.; Effect sizes
associated with Chi-squared and F distributions), this applies also to cases
where the lower bound of the CI is equal to 0. Even more care should be taken
when the \emph{upper} bound is equal to 0 - this occurs when \emph{p}-value is greater
than 1-alpha/2 making, the upper bound cannot be estimated, and the upper
bound is arbitrarily set to 0 (Steiger, 2004). For example:\if{html}{\out{<div class="r">}}\preformatted{eta_squared(aov(mpg ~ factor(gear) + factor(cyl), mtcars[1:7, ]))
}\if{html}{\out{</div>}}\preformatted{## # Effect Size for ANOVA (Type I)
## 
## Parameter    | Eta2 (partial) |       90\% CI
## --------------------------------------------
## factor(gear) |           0.58 | [0.00, 1.00]
## factor(cyl)  |           0.46 | [0.00, 1.00]
}
}

\section{CI Does Not Contain the Estimate}{

For very large sample sizes, the width of the CI can be smaller than the
tolerance of the optimizer, resulting in CIs of width 0. This can also,
result in the estimated CIs excluding the point estimate. For example:\if{html}{\out{<div class="r">}}\preformatted{t_to_d(80, df_error = 4555555)
}\if{html}{\out{</div>}}\preformatted{## d    |       95\% CI
## -------------------
## 0.07 | [0.08, 0.08]
}
}

\section{Choosing the Appropriate CI Level}{

"Confidence intervals on measures of effect size convey all the information
in a hypothesis test, and more" (Steiger, 2004). Essentially, a hypothesis
test can be preformed by inspecting the CI - if it excludes the null
hypothesized value, then we can conclude that the effect size is
significantly different from this value. For 2-sided tests, such as those
involving \emph{t}- or \emph{z}-statistics, the CI level should complement the
significance level. So if \eqn{\alpha} = 5\%, then the confidence level should
be 1-\eqn{\alpha} = 1-5\% = 95\%.
\cr\cr
However, one-sided hypothesis tests, such as those involving \emph{F}- or
\eqn{\chi^2}-statistics (or one-tailed \emph{t}- or \emph{z}-tests), a two sided CI can
include the null value even when the test itself is significant. One common
solution is to use a two sided CI with a confidence level of 1-2\eqn{\alpha}
= 1-2*5\% = 90\%. This is why across \code{effectsize}, effect sizes associated with
\emph{F}- or \eqn{\chi^2}-statistics (Cramer's \emph{V}, \eqn{\eta^2}, ...) default to
a 90\% CI.
\cr\cr
The problem with this solution is that it doesn't actually give 95\% coverage
on the estimated effect size. To this end we can use a one-sided 95\% CI. This
can be done by setting the upper bound of the CI to \code{Inf}. For example:\if{html}{\out{<div class="r">}}\preformatted{data("hardlyworking")
fit <- lm(salary ~ n_comps + age, data = hardlyworking)
etasq <- eta_squared(fit, ci = 0.90)
etasq[["CI"]] <- 0.95
etasq[["CI_high"]] <- Inf
etasq
}\if{html}{\out{</div>}}\preformatted{## # Effect Size for ANOVA (Type I)
## 
## Parameter | Eta2 (partial) |           95\% CI
## ---------------------------------------------
## n_comps   |           0.21 | [0.17,      Inf]
## age       |           0.10 | [0.07,      Inf]
}
}

\references{
\itemize{
\item Morey, R. D., Hoekstra, R., Rouder, J. N., Lee, M. D., & Wagenmakers, E. J. (2016). The fallacy of placing confidence in confidence intervals. Psychonomic bulletin & review, 23(1), 103-123.
\item Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis. Psychological Methods, 9, 164-182.
}
}
