% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{effectsize_deprecated}
\alias{effectsize_deprecated}
\alias{probs_to_odds.data.frame}
\alias{oods_to_probs.data.frame}
\title{Deprecated / Defunct Functions}
\arguments{
\item{...}{Arguments to the deprecated function.}
}
\description{
Deprecated / Defunct Functions
}
