% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{effectsize_deprecated}
\alias{effectsize_deprecated}
\alias{convert_odds_to_probs}
\alias{convert_probs_to_odds}
\alias{convert_d_to_r}
\alias{convert_r_to_d}
\alias{convert_oddsratio_to_d}
\alias{convert_d_to_oddsratio}
\alias{convert_oddsratio_to_r}
\alias{convert_r_to_oddsratio}
\alias{interpret_d}
\alias{interpret_g}
\alias{interpret_delta}
\alias{interpret_parameters}
\alias{normalized_chi}
\alias{chisq_to_normalized}
\alias{convert_d_to_common_language}
\alias{d_to_common_language}
\alias{convert_rb_to_common_language}
\alias{rb_to_common_language}
\alias{common_language}
\title{Deprecated / Defunct Functions}
\usage{
convert_odds_to_probs(...)

convert_probs_to_odds(...)

convert_d_to_r(...)

convert_r_to_d(...)

convert_oddsratio_to_d(...)

convert_d_to_oddsratio(...)

convert_oddsratio_to_r(...)

convert_r_to_oddsratio(...)

interpret_d(...)

interpret_g(...)

interpret_delta(...)

interpret_parameters(...)

normalized_chi(...)

chisq_to_normalized(...)

convert_d_to_common_language(...)

d_to_common_language(...)

convert_rb_to_common_language(...)

rb_to_common_language(...)

common_language(...)
}
\arguments{
\item{...}{Arguments to the deprecated function.}
}
\description{
Deprecated / Defunct Functions
}
