% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{effectsize_deprecated}
\alias{effectsize_deprecated}
\alias{interpret_d}
\alias{interpret_g}
\alias{interpret_delta}
\alias{interpret_parameters}
\alias{normalized_chi}
\alias{chisq_to_normalized}
\title{Deprecated / Defunct Functions}
\usage{
interpret_d(...)

interpret_g(...)

interpret_delta(...)

interpret_parameters(...)

normalized_chi(...)

chisq_to_normalized(...)
}
\arguments{
\item{...}{Arguments to the deprecated function.}
}
\description{
Deprecated / Defunct Functions
}
\details{
\itemize{
\item \code{interpret_d} is now \code{\link{interpret_cohens_d}}.
\item \code{interpret_g} is now \code{\link{interpret_hedges_g}}.
\item \code{interpret_delta} is now \code{\link{interpret_glass_delta}}.
\item \code{interpret_parameters} for \emph{standardized parameters} was incorrect. Use \code{\link{interpret_r}} instead.
\item \code{normalized_chi} is now \code{\link{fei}}.
\item \code{chisq_to_normalized} is now \code{\link{chisq_to_fei}}.
\item \code{d_to_cles} and \code{rb_to_cles} are now one of the available functions for CLES conversion, e.g. \code{\link{d_to_U1}}.
}
}
