% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_test.R
\name{equivalence_test.effectsize_table}
\alias{equivalence_test.effectsize_table}
\title{Equivalence test}
\usage{
\method{equivalence_test}{effectsize_table}(x, range = "default", ...)
}
\arguments{
\item{x}{An effect size table, such as returned by \code{cohens_d, eta_squared, F_to_r}, etc.}

\item{range}{The range of practical equivalence of an effect. If a single value is provided,
the test is done against \code{c(-range, range)}. For effect sizes that cannot be negative,
the lower bound is set to 0. If \code{"default"}, will be set to \code{[-.1, .1]}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame.
}
\description{
Compute the equivalence test for frequentist models.
}
\examples{

ds <- t_to_d(t = c(0.45, -0.65, -2.2, 2.25),
             df_error = c(675, 525, 900, 1875))
(equi <- equivalence_test(ds, range = 0.2))
if (require(see)) plot(equi)


model <- aov(mpg ~ factor(am) * factor(cyl), data = mtcars)
es <- eta_squared(model)
(equi <- equivalence_test(es, range = 0.15))
if (require(see)) plot(equi)

}
\seealso{
For more details, see \code{\link[bayestestR:equivalence_test]{equivalence_test}}.
}
