% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hardlyworking}
\alias{hardlyworking}
\title{Workers' salary and other information}
\format{
A data frame with 500 rows and 5 variables:
\describe{
\item{salary}{Salary, in Shmekels}
\item{xtra_hours}{Number of overtime hours (on average, per week)}
\item{n_comps}{Number of compliments given to the boss (observed over the last week)}
\item{age}{Age in years}
\item{seniority}{How many years with the company}
}
}
\usage{
hardlyworking
}
\description{
A sample (simulated) dataset, used in tests and some examples.
}
\keyword{data}
