% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret}
\alias{interpret}
\alias{interpret.numeric}
\title{Generic function for interpretation}
\usage{
interpret(x, ...)

\method{interpret}{numeric}(x, rules, name = attr(rules, "rule_name"), ...)
}
\arguments{
\item{x}{Vector of value break points (edges defining categories).}

\item{...}{Currently not used.}

\item{rules}{Set of \code{\link[=rules]{rules()}}.}

\item{name}{Name of the set of rules (stored as a 'rule_name' attribute).}
}
\description{
Interpret a value based on a set of rules. See \code{\link[=rules]{rules()}}.
}
\examples{
rules_grid <- rules(c(0.01, 0.05), c("very significant", "significant", "not significant"))
interpret(0.001, rules_grid)
interpret(0.021, rules_grid)
interpret(0.08, rules_grid)
interpret(c(0.01, 0.005, 0.08), rules_grid)

interpret(c(0.35, 0.15), c("small" = 0.2, "large" = 0.4), name = "Cohen's Rules")
interpret(c(0.35, 0.15), rules(c(0.2, 0.4), c("small", "medium", "large")))
}
\seealso{
rules
}
