% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_d.R
\name{interpret_d}
\alias{interpret_d}
\alias{interpret_g}
\alias{interpret_delta}
\title{Interpret standardized differences}
\usage{
interpret_d(d, rules = "cohen1988")

interpret_g(g, rules = "cohen1988")

interpret_delta(delta, rules = "cohen1988")
}
\arguments{
\item{d, g, delta}{Value or vector of effect size values.}

\item{rules}{Can be \code{"cohen1988"} (default), \code{"gignac2016"}, \code{"sawilowsky2009"} or custom set of \code{\link[=rules]{rules()}}.}
}
\description{
Interpretation of standardized differences using different sets of rules of thumb.
}
\examples{
interpret_d(.02)
interpret_d(c(.5, .02))

}
\references{
\itemize{
\item Gignac, G. E., & Szodorai, E. T. (2016). Effect size guidelines for individual differences researchers. Personality and individual differences, 102, 74-78.
\item Cohen, J. (1988). Statistical power analysis for the behavioural sciences.
\item Sawilowsky, S. S. (2009). New effect size rules of thumb.
}
}
