% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_bayesian_indices.R
\name{interpret_ess}
\alias{interpret_ess}
\alias{interpret_rhat}
\alias{interpret_rope}
\title{Bayesian indices interpretation}
\usage{
interpret_ess(ess, rules = "burkner2017")

interpret_rhat(rhat, rules = "vehtari2019")

interpret_rope(rope, ci = 0.9, rules = "default")
}
\arguments{
\item{ess}{Value or vector of Effective Sample Size (ESS) values.}

\item{rules}{A character string (see details) or a custom set of \code{\link[=rules]{rules()}}.}

\item{rhat}{Value or vector of Rhat values.}

\item{rope}{Value or vector of percentages in ROPE.}

\item{ci}{The Credible Interval (CI) probability, corresponding to the proportion of HDI, that was used. Can be \code{1} in the case of "full ROPE".}
}
\description{
Interpretation of Bayesian indices, such as Effective Sample Size (ESS), Rhat, or percentage in ROPE.
}
\details{
\subsection{Rules sets:}{
\itemize{
\item \strong{ESS}: Can be "burkner2017" (default).
\item \strong{Rhat}: Can be "vehtari2019" (default) or "gelman1992".
\item \strong{ROPE}: Can be \href{https://easystats.github.io/bayestestR/articles/guidelines.html}{"default"}.
\item \strong{ESS}:
}}
}
\examples{
interpret_ess(1001)
interpret_ess(c(852, 1200))

interpret_rhat(1.00)
interpret_rhat(c(1.5, 0.9))

interpret_rope(0, ci = 0.9)
interpret_rope(c(0.005, 0.99), ci = 1)
}
\references{
\itemize{
\item Bürkner, P. C. (2017). brms: An R package for Bayesian multilevel models using Stan. Journal of Statistical Software, 80(1), 1-28.
\item Gelman, A., & Rubin, D. B. (1992). Inference from iterative simulation using multiple sequences. Statistical science, 7(4), 457-472.
\item Vehtari, A., Gelman, A., Simpson, D., Carpenter, B., & Bürkner, P. C. (2019). Rank-normalization, folding, and localization: An improved Rhat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.
\item \href{https://easystats.github.io/bayestestR/articles/guidelines.html}{BayestestR's reporting guidelines}
}
}
