% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_kendalls_w.R
\name{interpret_kendalls_w}
\alias{interpret_kendalls_w}
\title{Interpret Kendall's coefficient of concordance}
\usage{
interpret_kendalls_w(w, rules = "landis1977")
}
\arguments{
\item{w}{Value or vector of Kendall's coefficient of concordance.}

\item{rules}{Can be \code{"landis1977"} (default) or a custom set of \code{\link[=rules]{rules()}}.}
}
\description{
Interpret Kendall's coefficient of concordance
}
\section{Rules}{


Landis & Koch (1977) (\code{"landis1977"}):
\itemize{
\item \strong{w <= 0.2} ~ slight agreement
\item \strong{0.2 < w <= 0.4} ~ fair agreement
\item \strong{0.4 < w <= 0.6} ~ moderate agreement
\item \strong{0.6 < w <= 0.8} ~ substantial agreement
\item \strong{0.8 < w} ~ almost perfect agreement
}
}

\references{
\itemize{
\item Landis, J. R., & Koch G. G. (1977). The measurement of observer agreement
for categorical data. Biometrics, 33:159-74.
}
}
