% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_effectsize_name.R
\name{is_effectsize_name}
\alias{is_effectsize_name}
\alias{get_effectsize_label}
\title{Checks if character is of a supported effect size}
\usage{
is_effectsize_name(x, ignore_case = TRUE)

get_effectsize_label(x, ignore_case = TRUE)
}
\arguments{
\item{x}{A character, or a vector.}

\item{ignore_case}{Should case of input be ignored?}
}
\description{
For use by other functions and packages.
}
