% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_between_riskdiff.R
\name{oddsratio_to_riskratio}
\alias{oddsratio_to_riskratio}
\alias{riskratio_to_oddsratio}
\alias{riskratio_to_logoddsratio}
\alias{riskratio_to_arr}
\alias{oddsratio_to_arr}
\alias{logoddsratio_to_arr}
\alias{arr_to_riskratio}
\alias{arr_to_oddsratio}
\alias{arr_to_logoddsratio}
\alias{arr_to_nnt}
\alias{nnt_to_arr}
\title{Convert Between Odds Ratios and Risk Ratios}
\usage{
oddsratio_to_riskratio(OR, p0, log = FALSE, verbose = TRUE, ...)

riskratio_to_oddsratio(RR, p0, log = FALSE, verbose = TRUE, ...)

riskratio_to_logoddsratio(RR, p0, log = TRUE, verbose = TRUE, ...)

riskratio_to_arr(RR, p0, log = FALSE, verbose = TRUE, ...)

oddsratio_to_arr(OR, p0, log = FALSE, verbose = TRUE, ...)

logoddsratio_to_arr(OR, p0, log = TRUE, verbose = TRUE, ...)

arr_to_riskratio(ARR, p0, log = FALSE, verbose = TRUE, ...)

arr_to_oddsratio(ARR, p0, log = FALSE, verbose = TRUE, ...)

arr_to_logoddsratio(ARR, p0, log = TRUE, verbose = TRUE, ...)

arr_to_nnt(x)

nnt_to_arr(x)
}
\arguments{
\item{OR, RR, ARR}{Risk ratio of \code{p1/p0} Odds ratio of \code{odds(p1)/odds(p0)}
(possibly log-ed), or Absolute Risk Reduction or \code{p1 - p0}. \code{OR} can also
be a logistic regression model.}

\item{p0}{Baseline risk}

\item{log}{Take in or output the log of the ratio (such as in logistic models).}

\item{verbose}{Toggle warnings and messages on or off.}

\item{...}{Arguments passed to and from other methods.}

\item{x}{Absolute Risk Reduction or Number Needed to Treat.}
}
\value{
Converted index, or if \code{OR} is a logistic regression model, a
parameter table with the converted indices.
}
\description{
Convert Between Odds Ratios and Risk Ratios
}
\examples{
p0 <- 0.4
p1 <- 0.7

(OR <- probs_to_odds(p1) / probs_to_odds(p0))
(RR <- p1 / p0)
(ARR <- p1 - p0)
(NNT <- arr_to_nnt(ARR))

riskratio_to_oddsratio(RR, p0 = p0)
oddsratio_to_riskratio(OR, p0 = p0)
riskratio_to_arr(RR, p0 = p0)
arr_to_oddsratio(nnt_to_arr(NNT), p0 = p0)

m <- glm(am ~ factor(cyl),
  data = mtcars,
  family = binomial()
)
oddsratio_to_riskratio(m, verbose = FALSE) # RR is relative to the intercept if p0 not provided

}
\references{
Grant, R. L. (2014). Converting an odds ratio to a range of plausible
relative risks for better communication of research findings. Bmj, 348,
f7450.
}
\seealso{
\code{\link[=oddsratio]{oddsratio()}}, \code{\link[=riskratio]{riskratio()}}, \code{\link[=arr]{arr()}}, and \code{\link[=nnt]{nnt()}}.

Other convert between effect sizes: 
\code{\link{d_to_r}()},
\code{\link{diff_to_cles}},
\code{\link{eta2_to_f2}()},
\code{\link{odds_to_probs}()}
}
\concept{convert between effect sizes}
