% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_cramers_v.R
\name{phi}
\alias{phi}
\alias{cramers_v}
\title{Effect size for contingency tables}
\usage{
phi(x, y = NULL, ...)

cramers_v(x, y = NULL, ...)
}
\arguments{
\item{x}{a numeric vector or matrix. x and y can also both be factors.}

\item{y}{a numeric vector; ignored if x is a matrix. If x is a factor, y should be a factor of the same length.}

\item{...}{Passed to \code{chisq.test()}.}
}
\value{
A numeric value between 0-1.
}
\description{
Compute Cramer's V and phi (\eqn{\phi}) from contingency tables.
}
\details{
These functions use the following formulae:
\cr\cr
\deqn{Cramer's V = \sqrt{\chi^2 / (n * (min(nrow,ncol)-1))}}
\cr\cr
\deqn{\phi = \sqrt{\chi^2 / n}}
}
\examples{
contingency_table <- as.table(rbind(c(762, 327, 468), c(484, 239, 477), c(484, 239, 477)))

phi(contingency_table)

cramers_v(contingency_table)
}
