% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_semipartial.R
\name{r2_semipartial}
\alias{r2_semipartial}
\title{Semi-Partial Correlation Squared (aka Delta R2)}
\usage{
r2_semipartial(
  model,
  type = c("terms", "parameters"),
  ci = 0.95,
  alternative = "greater",
  ...
)
}
\arguments{
\item{model}{An \code{lm} model.}

\item{type}{Type, either "term", or "parameters".}

\item{ci}{Confidence interval, defaults to 95\%.}

\item{alternative}{Alternative hypothesis for the confidence
interval. One of “two.sided”, “less”, or “greater” (default).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame with the effect size.
}
\description{
Compute the semi-partial correlation squared (also known as the delta
R2), for an \code{lm} model.
}
\details{
This is similar to the last column of the "Conditional Dominance
Statistics" section of the \code{\link[parameters:dominance_analysis]{parameters::dominance_analysis()}} output.
For other, non-lm models, as well as more verbose info / options,
please see the documentation for \code{\link[parameters:dominance_analysis]{parameters::dominance_analysis()}}.
}
\examples{
m <- lm(mpg ~ factor(cyl) + disp + hp * drat, data = mtcars)
r2_semipartial(m)
}
\concept{effect size correlation}
