% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{equivalence_test}
\alias{standardize}
\alias{standardise}
\alias{standardize_parameters}
\alias{standardize_posteriors}
\alias{standardize_info}
\alias{display}
\alias{print_html}
\alias{print_md}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{equivalence_test}}}

  \item{datawizard}{\code{\link[datawizard:standardize]{standardise}}, \code{\link[datawizard]{standardize}}}

  \item{insight}{\code{\link[insight]{display}}, \code{\link[insight:display]{print_html}}, \code{\link[insight:display]{print_md}}}

  \item{parameters}{\code{\link[parameters]{standardize_info}}, \code{\link[parameters]{standardize_parameters}}, \code{\link[parameters:standardize_parameters]{standardize_posteriors}}}
}}

