% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_pooled.R
\name{sd_pooled}
\alias{sd_pooled}
\alias{mad_pooled}
\title{Pooled Standard Deviation}
\usage{
sd_pooled(x, y = NULL, data = NULL)

mad_pooled(x, y = NULL, data = NULL)
}
\arguments{
\item{x}{A continuous variable or a formula.}

\item{y}{A continuous variable, a factor with two groups or a formula.}

\item{data}{An optional data frame containing the variables.}
}
\value{
Numeric, the pooled standard deviation.
}
\description{
The Pooled Standard Deviation is a weighted average of standard deviations for two or more groups, with more "weight" given to larger sample sizes.
}
\examples{
sd_pooled(Sepal.Length ~ Petal.Width, data = iris)
}
