% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_pooled.R
\name{sd_pooled}
\alias{sd_pooled}
\alias{mad_pooled}
\title{Pooled Standard Deviation}
\usage{
sd_pooled(x, y = NULL, data = NULL)

mad_pooled(x, y = NULL, data = NULL)
}
\arguments{
\item{x}{A formula, a numeric vector, or a name of one in \code{data}.}

\item{y}{A numeric vector, a grouping (character / factor) vector, a or a name of one in \code{data}. Ignored if \code{x} is a formula.}

\item{data}{An optional data frame containing the variables.}
}
\value{
Numeric, the pooled standard deviation.
}
\description{
The Pooled Standard Deviation is a weighted average of standard deviations for two or more groups, with more "weight" given to larger sample sizes.
}
\examples{
sd_pooled(mpg ~ am, data = mtcars)

}
